Improve powdr lib (#1980)
Just make the Session stuff a bit cleaner.
The removed files were merged by accident.

+++ riscv/src/large_field/code_gen.rs
@@ -211,13 +211,13 @@ let initial_memory: (fe, fe)[] = [
         runtime.submachines_import(),
         1 << (options
             .min_degree_log
-            .unwrap_or(powdr_linker::MIN_DEGREE_LOG as u32)),
+            .unwrap_or(powdr_linker::MIN_DEGREE_LOG as u8)),
         // We expect some machines (e.g. register memory) to use up to 4x the number
         // of rows as main. By setting the max degree of main to be smaller by a factor
         // of 4, we ensure that we don't run out of rows in those machines.
         1 << options
             .max_degree_log
-            .unwrap_or(*powdr_linker::MAX_DEGREE_LOG as u32 - 2),
+            .unwrap_or(*powdr_linker::MAX_DEGREE_LOG as u8 - 2),
         runtime.submachines_declare(),
         preamble,
         initial_memory
